+++ Kindlesort/views.py
@@ -1,7 +1,9 @@
 import os 
-from flask import Flask, flash, request, redirect, url_for, render_template
-from Kindlesort import app, ALLOWED_EXTENSIONS
+from flask import Flask, flash, request, redirect, url_for, render_template, session
+from Kindlesort import app, ALLOWED_EXTENSIONS, SECRET_KEY
 from werkzeug.utils import secure_filename
+from myClippings import get_clippings
+
 
 def allowed_file(filename): 
     return '.' in filename and \
@@ -17,14 +19,11 @@ def index():
     '''The home template'''
     return render_template("home.html")
 
-@app.route("/books", methods=['GET', 'POST'])  
-def books(): 
+@app.route("/upload", methods=['GET', 'POST'])  
+def upload(): 
     '''The template containing books, ability to upload myclippings'''
     if request.method == 'POST':
         # check if the post request has the file part
-        if 'file' not in request.files:
-            flash('No file part')
-            return redirect(request.url)
         file = request.files['file']
         # If the user does not select a file, the browser submits an
         # empty file without a filename.
@@ -34,8 +33,18 @@ def books():
         if file and allowed_file(file.filename):
             filename = secure_filename(file.filename)
             file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
-        return render_template('books.html')
 
 
+        return render_template('upload.html')
+
+
+
+    return render_template("upload.html")
+
+@app.route("/books", methods=['GET', 'POST'])
+def books(): 
+    if request.method == 'POST': 
+        if 'submit_button' in request.form: 
+            return render_template("books.html", books=get_clippings())
 
     return render_template("books.html")