+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 7/30/22
+- Seperated books page from upload page 
+- Added myClippings.py 
+- Added list objects of all the clippings text in the books page
+
 # 7/29/22
 - Started adding bootstrap
 - Made base.html so I can extend it as a template to other templates